update logs & version
